Remove unnecessary toString casts and move TagCriteria to main package
